pyproject.toml
[tool.poetry]
name = "oasis-tg-bot"
version = "0.1.0"
description = "Telegram bot for analyzing forwarded messages for potential account compromise"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^20.7"
openai = "^1.3.0"
aiohttp = "^3.9.0"
dotenv = "^0.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
oasis-tg-bot = "bot:main" 